StudyHub - BackEnd/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
python-jose
passlib[bcrypt]
pydantic
python-multipart
aiofiles